Resolve a mountain of warnings related to String removal and casting
a small amount of this was not actually directly involved in the String removal, but are necessary based on the changes to the Str functions. it would actually be better to remove some of the Str functions. but that would require effort

+++ src/System/File.cpp
@@ -189,11 +189,11 @@ static std::string Concatenate(const std::string& first, const std::string& seco
 
 	// Reconstruct the rest of the path from the items.
 	auto i = items.begin();
-	Str::append(out, i->p, static_cast<int>(i->n));
+	Str::append(out, i->p, i->n);
 	for(++i; i != items.end(); ++i)
 	{
 		Str::append(out, '\\');
-		Str::append(out, i->p, static_cast<int>(i->n));
+		Str::append(out, i->p, i->n);
 	}
 
 	// End with a slash if requested.
@@ -349,7 +349,7 @@ std::string Path::brief() const
 	std::string out = top();
 	if(out.length() > 20)
 	{
-		Str::erase(out, 3, static_cast<int>(out.length()) - 16);
+		Str::erase(out, 3, out.length() - 16);
 		Str::insert(out, 3, '~');
 	}
 	return out;
@@ -528,7 +528,7 @@ Vector<std::string> getLines(const std::string& path, bool* success)
 			}
 			if(end > pos)
 			{
-				Str::append(out.back(), buffer.data() + pos, static_cast<int>(end - pos));
+				Str::append(out.back(), buffer.data() + pos, end - pos);
 			}
 			if(end < bytesRead && isNewline(buffer[end]))
 			{
